Extracted the article banner into its own component
