+++ src/components/Header/styles.module.scss
@@ -15,7 +15,7 @@
   .logo {
     color: $greenFern;
     cursor: pointer;
-    font-family: $Titillium;
+    font-family: $titillium;
     font-size: 30px;
     font-weight: bold;
   }
